Backend Chatbot/event-spectra-backend-python/requirements.txt
Flask
Flask-Cors
Werkzeug
langchain
openai
python-dotenv
langchain-openai 
Gunicorn
langchain_experimental
pandas
tabulate